+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 ## General
 
 * Improved Kubernetes definition validation
+* Allow RKE2 deployments with Calico, Cilium and Multus on aarch64 platforms
 
 ## API
 
@@ -13,6 +14,8 @@
 ### Image Configuration Directory Changes
 
 ## Bug Fixes
+
+* [#491](https://github.com/suse-edge/edge-image-builder/issues/491) - Large Helm manifests fail to install
 
 ---